src/XYW-Net_original/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
opencv-python>=4.5.0
matplotlib>=3.3.0
scipy>=1.5.0
pillow>=8.0.0
pyyaml>=5.3.0
tqdm>=4.50.0
tensorboard>=2.4.0
thop>=0.1.1
scikit-image>=0.17.0
requests>=2.25.0
